REFACTOR :: token response 생성
