+++ src/user/dto/response/token-response.dto.ts
@@ -1,0 +1,10 @@
+import { Expose, plainToInstance } from 'class-transformer';
+
+export class TokenResponseDto {
+  @Expose()
+  token: string;
+
+  static of(token: string): TokenResponseDto {
+    return plainToInstance(TokenResponseDto, token);
+  }
+}
